+++ features/refinance/helpers/replaceETHwithWETH.ts
@@ -1,6 +1,12 @@
-export const replaceETHWithWETH = (symbol: string) => {
-  if (symbol === 'ETH') {
-    return 'WETH'
+import { mapTokenToSdkToken } from 'features/refinance/mapTokenToSdkToken'
+import { TokenAmount } from 'summerfi-sdk-common'
+
+export const replaceETHWithWETH = (tokenAmount: TokenAmount) => {
+  if (tokenAmount.token.symbol === 'ETH') {
+    return TokenAmount.createFrom({
+      token: mapTokenToSdkToken(tokenAmount.token.chainInfo, 'WETH'),
+      amount: tokenAmount.amount,
+    })
   }
-  return symbol
+  return tokenAmount
 }
